Make the Dimension class assume integers, as the tests seem to indicate they shoul dbe (rather than doubles)

+++ Source/TitaniumKit/include/Titanium/UI/Dimension.hpp
@@ -24,10 +24,10 @@ namespace Titanium
 		  See http://docs.appcelerator.com/titanium/latest/#!/api/Dimension
 		  */
 		struct Dimension {
-			double x { 0 };
-			double y { 0 };
-			double width { 0 };
-			double height { 0 };
+			int32_t x { 0 };
+			int32_t y { 0 };
+			uint32_t width { 0 };
+			uint32_t height { 0 };
 		};
 
 		Dimension js_to_Dimension(const JSObject& object);
